pkvm: x86: Add pkvm_hyp_types.h for pkvm_constants
pkvm_constants.c uses hyp/pkvm_hyp.h (which is generally supposed to be
used by the hyp code only) in order to know sizes of some of the hyp
structs. As a result, hyp/pkvm_hyp.h cannot include definitions (e.g.
static inline functions) that refer to symbols that are not visible to
the host code.

To overcome this limitation, move those struct definitions used by
pkvm_constants.c to a separate header hyp/pkvm_hyp_types.h, so that
pkvm_constants.c only includes the definitions it actually needs.

This is a preparation for adding gpa_range_has_pvmfw() inline function
to hyp/pkvm_hyp.h.
